requirements.txt
requests==2.31.0
asyncpg==0.28.0
databases[postgresql]==0.2.6
fastapi==0.48.0
SQLAlchemy==1.3.13
httpx==0.11.1
python-decouple==3.8
validate_email_address
py3dns